feature(s3): s3 이미지 업로드 삭제 구현
